+++ CHANGELOG.md
@@ -11,11 +11,17 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Pipelines which use multiple descriptor sets (different `set =` values) sometimes trigger
   validation errors
+- `contrib/screen-13-hot`: build error on Windows platform
 
 ### Added
 
+- Custom vertex layout support - see `examples/vertex_layout.rs`
 - Enabled full set of Vulkan 1.1 and Vulkan 1.2 core features during device creation
 - Ray query support with `ray_omni.rs` example
+- Exposed existing command buffer implementation so that programs may wait for render graph GPU
+  submissions to finish executing before reading the results with the CPU - see
+  `examples/cpu_readback.rs`
+- `KeyBuf::is_down` helper function
 
 ### Changed